skip verify tls certs when publishing server desc
The current directory server - anemone.mooo.com - uses
a self-signed certificate and fails verification. As Zax
says, this is a kludgy fix, but the verification is not
necessary because the ServerDesc's are signed by the
nodes. With this fix, more modern versions of python 2.7
will be able to upload server descs.